refactor: 유저 API 리팩터링 (#114)
